requirements.txt
# requirements.txt
torch>=1.9.0
torchvision>=0.10.0
pyyaml>=5.0
tqdm>=4.60.0
tensorboard>=2.5.0
scikit-learn>=0.24.0
pytest>=6.2.0